chore: modularize django common app
