+++ bc_obps/common/models/__init__.py
@@ -1,0 +1,12 @@
+# Order of imports is important to avoid circular dependencies
+from .base_model import BaseModel
+from .dashboard_data import DashboardData
+from .email_notification import EmailNotification
+from .email_notification_template import EmailNotificationTemplate
+
+__all__ = [
+    "BaseModel",
+    "DashboardData",
+    "EmailNotification",
+    "EmailNotificationTemplate",
+]
